cleanup

+++ susumu/src/platform/windows/windows_input.cpp
@@ -9,14 +9,14 @@ namespace susumu
     bool Input::IsKeyPressed(KeyCode key)
     {
         auto window = static_cast<GLFWwindow*>(App::Get().GetWindow().GetNativeWindow());
-        auto state = glfwGetKey(window, static_cast<int32_t>(key));
+        auto state = glfwGetKey(window, key);
         return state == GLFW_PRESS || state == GLFW_REPEAT;
     }
 
     bool Input::IsMouseButtonPressed(MouseCode button)
     {
         auto window = static_cast<GLFWwindow*>(App::Get().GetWindow().GetNativeWindow());
-        auto state = glfwGetMouseButton(window, static_cast<int32_t>(button));
+        auto state = glfwGetMouseButton(window, button);
         return state == GLFW_PRESS;
     }
